code/requirements.txt
antlr4-python3-runtime==4.9.3
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
commonroad-drivability-checker==2022.1
commonroad-io==2022.1
git+https://gitlab.lrz.de/tum-cps/commonroad-route-planner.git@271e13b51cce8ea5494d9f13a611f46a57207915#egg=commonroad_route_planner
commonroad-vehicle-models==2.0.0
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
flask-marshmallow==0.14.0
fonttools==4.38.0
idna==3.4
iniconfig==1.1.1
iso3166==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.18.0
matplotlib==3.3.4
networkx==2.8.7
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
Polygon3==3.0.9.1
protobuf==3.20.1
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-timeout==2.1.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
Rtree==1.0.1
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.15
Werkzeug==2.2.2

